Lock to EST
I tried `moment().utcOffset()` and `new Date().getTimezoneOffset()` but
could not get the user's timezone. Maybe because it's always running in
AWS Lambda. There must be some way to get the timezone from Alexa. Or not:
https://forums.developer.amazon.com/questions/29445/any-way-to-get-users-timezone.html
But I can ask Alexa "What time zone am i in?" and get a correct response.